dev/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.6
# To update, run:
#
#    pip-compile
#
asgiref==3.4.1
    # via django
certifi==2021.5.30
    # via requests
cffi==1.14.5
    # via cryptography
chardet==4.0.0
    # via requests
click==8.0.1
    # via pip-tools
closure==20191111
    # via uil-django-core
cryptography==3.4.7
    # via -r requirements.in
django==3.2.5
    # via
    #   -r requirements.in
    #   django-braces
    #   django-csp
    #   django-debug-toolbar
    #   django-extensions
    #   django-modeltranslation
    #   django-simple-menu
    #   djangorestframework
    #   uil-django-core
django-braces==1.14.0
    # via -r requirements.in
django-csp==3.7
    # via -r requirements.in
django-debug-toolbar==3.2.1
    # via -r requirements.in
django-extensions==3.1.3
    # via -r requirements.in
django-impersonate==1.7.3
    # via -r requirements.in
django-modeltranslation==0.17.3
    # via -r requirements.in
django-simple-menu==1.2.1
    # via -r requirements.in
djangorestframework==3.12.4
    # via uil-django-core
idna==2.10
    # via requests
importlib-metadata==4.6.1
    # via
    #   click
    #   pep517
lesscpy==0.14.0
    # via uil-django-core
mysqlclient==2.0.3
    # via -r requirements.in
pep517==0.10.0
    # via pip-tools
pip-tools==6.2.0
    # via -r requirements.in
ply==3.11
    # via lesscpy
pscript==0.7.5
    # via vbuild
pycparser==2.20
    # via cffi
pyjwt==2.1.0
    # via uil-django-core
pyscss==1.3.7
    # via uil-django-core
pytz==2021.1
    # via django
requests==2.25.1
    # via uil-django-core
six==1.16.0
    # via
    #   django-braces
    #   django-modeltranslation
    #   lesscpy
    #   pyscss
sqlparse==0.4.1
    # via
    #   django
    #   django-debug-toolbar
toml==0.10.2
    # via pep517
typing-extensions==3.10.0.0
    # via
    #   asgiref
    #   importlib-metadata
urllib3==1.26.6
    # via requests
vbuild==0.8.1
    # via uil-django-core
wheel==0.36.2
    # via pip-tools
zipp==3.5.0
    # via
    #   importlib-metadata
    #   pep517

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
